requirements.txt
Flask==2.3.2
pytest==7.1.3
pact-python==1.7.0
